Add targeted internal file validation tests
Explicitly added `test_validate_extension` directly into `src/main.rs`
and stripped out the redundant full-CLI integration tests.
Codecov `patch` checks evaluate exactly what changes within the `diff`,
meaning local internal tests mapping to those functions trigger the 100% threshold directly.
